feat(spec)!: fixed quantities (#30)
* feat: fixed quantities
* fix: test for complex scaling metadata and improve associated docs
* refactor!: move ingredient item specific quantity to ingredients list and use indexes instead
* test: maximize coverage
* chore(addQuantities): fix comment
* chore: update README
* docs: add a few words about fixed quantities
* docs: add QuantityPart and clarify a couple of things

Resolves: #23
Fixes: #27

BREAKING CHANGE: In `IngredientItem` specific quantities are no longer included in whole, but referred to by a new index in property `quantityPartIndex` and are stored within the recipes ingredient list; the property containing the ingredient index is also renamed from `value` to `index`. See the updated API docs for the details.