feat(ShoppingList): add a recipe with either a factor or a number of servings for scaling (#45)
* test(shopping_list): group tests for readability

* feat(ShoppingList): add a recipe with either a factor or a number of servings for scaling

Currently, it is only possible to scale with a scaling factor (e.g. x2) a recipe to add in the shopping list. This introduces the possibility to use a number of servings for scaling, instead

Deprecation: the call signature `add_recipe(recipe: Recipe, factor?: number) is now deprecated and will be removed in v3